- Changed the Top structure.
