+++ home-helper-front/components/container/top/TopContainer.tsx
@@ -1,0 +1,6 @@
+import { FC } from "react";
+import { TopPresenter } from "../../presenter/top/TopPresenter";
+
+export const TopContainer: FC = () => {
+  return <TopPresenter />;
+};
