+++ home-helper-front/components/page/top/TopPage.tsx
@@ -1,0 +1,6 @@
+import { FC } from "react";
+import { TopContainer } from "../../container/top/TopContainer";
+
+export const Page: FC = () => {
+  return <TopContainer />;
+};
