+++ home-helper-front/components/presenter/top/TopPresenter.tsx
@@ -1,0 +1,22 @@
+import { FC } from "react";
+import { NonHeaderLayout } from "../Layout";
+import MainTitle from "../../ui/MainTitle";
+import { CardLinkList } from "../../ui/Card";
+
+const pageList = [
+  { href: "/homeHelper", label: "お手伝いアプリ" },
+  { href: "/household", label: "家計簿アプリ" },
+];
+
+export const TopPresenter: FC = () => {
+  return (
+    <NonHeaderLayout
+      main={
+        <>
+          <MainTitle label={"うちのアプリへようこそ！"} />
+          <CardLinkList pageList={pageList} />
+        </>
+      }
+    />
+  );
+};
