+++ home-helper-front/pages/index.tsx
@@ -1,22 +1,6 @@
-import { CardLinkList } from "../components/ui/Card";
-import { NonHeaderLayout } from "../components/presenter/Layout";
-import MainTitle from "../components/ui/MainTitle";
-
-const pageList = [
-  { href: "/homeHelper", label: "お手伝いアプリ" },
-  { href: "/household", label: "家計簿アプリ" },
-];
+import { Page } from "../components/page/top/TopPage";
 
 const index = () => {
-  return (
-    <NonHeaderLayout
-      main={
-        <>
-          <MainTitle label={"うちのアプリへようこそ！"} />
-          <CardLinkList pageList={pageList} />
-        </>
-      }
-    />
-  );
+  return <Page />;
 };
 export default index;
